rearrange tests file structure, refactor fixtures to use specific `Collection`s rather than `Client`s
