+++ ci/stop_weaviate.sh
@@ -7,5 +7,5 @@ docker-compose -f ci/docker-compose-azure.yml down --remove-orphans
 docker-compose -f ci/docker-compose-okta-cc.yml down --remove-orphans
 docker-compose -f ci/docker-compose-okta-users.yml down --remove-orphans
 docker-compose -f ci/docker-compose-wcs.yml down --remove-orphans
-docker-compose -f ci/docker-compose-openai.yml down --remove-orphans
+docker-compose -f ci/docker-compose-generative.yml down --remove-orphans
 docker-compose -f ci/docker-compose-cluster.yml down --remove-orphans
